for test

+++ be/src/main/java/com/assignment/demo/repository/QuestionMaterialRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuestionMaterialRepository extends JpaRepository<QuestionMaterial,Integer> {
+public interface QuestionMaterialRepository extends JpaRepository<QuestionMaterial,Long> {
 }
